(WIP) Allocate buffers based on correct chroma format
The encoder currently allocates frame buffers before
it establishes what the chroma sub-sampling factor is,
always allocating based on the 4:4:4 format.

This patch detects the chroma format as early as
possible allowing the encoder to allocate buffers of
the correct size.

A future patch will change the encoder to allocate
frame buffers on demand to further reduce the memory
profile of the encoder.

Change-Id: Ic548ea6137acf0e8eb382f4540631ebbecff5bb8