Added high precision transforms
The high precision are only used if
configured with --enable-high-transforms

It gives greater precision in the transform.
This gives PSNR improvements when encoding
true 10 and 12 bit streams.

At the moment, the quantizer used is shifted
up by 2/4 for 10/12 bits so that the quantized
coefficients fit in the current token range.

Change-Id: Ia9c19a417cf030b8a7a889fcb3f5788bfca8215f