skip un-neccessary motion search in the first pass
This patch allows the encoder to skip the
un-neccessary motion search in the first pass. It
calculate the error of the zero motion vector using
the last source frame as reference and skip the
further motion search if the error is small.

Borg test shows little PSNR performance loss
(less than 1%). The first pass encoding speedup for
slideshow videos is over 30%.

Change-Id: I08b2ab110b695e4689573b2567fa531b6457616e