Add vp8 control function to reset the buffer level.
Add codec control function to allow for resetting, under change_config call,
the buffer level to the optimal_buffer_level (defined by the new bitrate).
The usage of this is for applications where the encoder is updated with a sudden change
in bitrate, and a faster response to the new bitrate is desired
(i.e., target_frame_size closer to per-frame-bandwidth).

Change-Id: I677e8b2e23b59a18a925f9f973de0f2ba587eaf2