Experiment for mid group second arf.
This patch implements a mechanism for inserting a second
arf at the mid position of arf groups.

It is currently disabled by default using the flag multi_arf_enabled.

Results are currently down somewhat in initial testing if
multi-arf is enabled. Most of the loss is attributable to the
fact that code to preserve the previous golden frame
(in the arf buffer) in cases where we are coding an overlay
frame, is currently disabled in the multi-arf case.

TBD in subsequent patches.
Remove old CONFIG_MULTIPLE_ARF structures.
Tuning of bit spend.
Some further clean up and refactoring.
Re-examine buffer flip options on overlay frames.

Change-Id: I1d777318ca09f147db2e8c86d7315fe86168c865