Fix test that checks only for keyframe
The first decoded frame does not have to be a keyframe,
it could be an inter-frame that is coded intra-only.

This patch fixes the test that checks what type the
frame is. Both keyframes and intra-only frames
contain the frame dimension in their frame header.

This patch also adds a unit test to check that
intra-only frames can be encoded as the very first
frames in the clip.

Change-Id: Ib40b1dbf91aae2bc047e23c626eaef09d1860147