Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Feels strange to minus 1 our own code, but I agree with the criticism that this is making the code unnecessarily complicated.  (In addition, it is arguable that the PSNR is being incorrectly computed in the downshift case as it is based on the downshifted image rather than the original source image.)

We'll adjust the patch so that only the same or upshifts in bit depth are allowed.

Patch-set: 3
Reviewer: Gerrit User 1115299 <1115299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0