skip the un-necessary motion search in the first pass
This patch allows the VP9 encoder to skip the un-necessary
motion search in the first pass. It computes the motion error
of 0,0 motion using the last source frame as the reference,
and skips the further motion search if this error is small.

Borg test shows overall there is little PSNR performance loss
(derf -0.001, yt 0.341, hd 0.282). Individual clips may have
PSNR gain or loss. The best PSNR performance is 7.347 and the
worst is -0.662. The first pass encoding speedup for slideshow
clips is over 30%.

Change-Id: I4cac4dbd911f277ee858e161f3ca652c771344fe