Update patch set 5
Patch Set 5: Looks good to me, approved

This is a security violation, so I think we should get it in as quickly as possible.    

Agree with james it would be nice to have a more comprehensive test of these, but I think that can wait for another day so we can land this.

Patch-set: 5
Reviewer: Gerrit User 1000963 <1000963@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2
Label: Commit-Queue=0
Label: Verified=0