Update patch set 12
Patch Set 12:

I included the 'break;' statements pointed out in an earlier changeset.

My original thinking was that this commit, up until that point, did not change the program functionality, so a change that did change functionality should be kept separate. However, it didn't occur to me that I could simply change the commit message to reflect the change, which removes that problem.

Patch-set: 12
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0