Intra-frame only test vector generation
This code generates a test vector for the
intra-only coded inter-frame.

It encodes three intra-only coded non-showable
frames, followed by two inter-coded frames,
followed by three "show existing frames" that
display the first three encoded frames in
reverse order, followed by two normally
encoded frames.

The frames themselves have a 'timecode' added
as an overlay to the video, so that it should be
easy to verify the correct displayed order by eye.
The correct order is as follows:
    3 4 2 1 0 8 9
input frames 5-7 are not displayed, or encoded,
instead 'display existing frame' is used to display
frames 2, 1, 0 in that order.

This code generates the test vector:
  vp90-2-16-intra-only.webm

when called with the following command line:
./vpxenc  y4m/bus_with_frame_num.y4m
  -o vp90-2-16-intra-only.webm
  --good --profile=0 --cpu-used=5
  --end-usage=vbr
  --min-q=0 --max-q=63
  --kf-min-dist=0 --kf-max-dist=150
  --auto-alt-ref=1
  --arnr-maxframes=7 --arnr-strength=5 --arnr-type=3
  --lag-in-frames=0
  --drop-frame=0 --static-thresh=0
  --bias-pct=50 --minsection-pct=0 --maxsection-pct=2000
  --sharpness=0
  --threads=0
  --undershoot-pct=100
  --test-decode=warn
  --resize-allowed=0
  --resize-width=1
  --resize-height=1
  --codec=vp9
  --limit=10
  --passes=1
  --target-bitrate=10000

Change-Id: Ieee83b56c1fa6d60525694f73d41684c0224b1c9