Intra-frame only test vector generation
This code generates a test vector for the
intra-only coded inter-frame.

It encodes three intra-only coded non-showable
frames, followed by two inter-coded frames,
followed by three "show existing frames" that
display the first three encoded frames in
reverse order, followed by two normally
encoded frames.

The frames themselves have a 'timecode' added
as an overlay to the video, so that it should be
easy to verify the correct displayed order by eye.
The correct order is as follows:
    3 4 2 1 0 8 9
input frames 5-7 are not displayed, or encoded,
instead 'display existing frame' is used to display
frames 2, 1, 0 in that order.

This code generates the test vector:
  vp90-2-16-intra-only.webm

Change-Id: Ieee83b56c1fa6d60525694f73d41684c0224b1c9