Update patch set 1
Patch Set 1:

When the coded was added in [1] it compared to cpu_used < -4, indicating real time, and the nonrd decision path didn't even exist yet. It wasn't until [2] where we eliminated negative speed levels that it applied to the rarely used good speed 5. I think that was an unintended side effect. 

[1] http://git.chromium.org/gitweb/?p=webm/libvpx.git;a=commitdiff;h=9e190dfb 
[2] http://git.chromium.org/gitweb/?p=webm/libvpx.git;a=commitdiff;h=56835f1b

Patch-set: 1