Make 2pass active_worst_quality adaptive
Before this commit, active_worst_quality was set once at the beginning
of the second pass, based on the average error from the first pass.
If this guess was too low, it was possible to completely miss the
target bitrate because the ideal quantizer of almost every frame was
bigger than active_worst_quality.

$ vpxenc --cpu-used=5  --fps=24/1 --target-bitrate=5000 --limit=300
         etv5k.y4m -o out.webm
=> Before: 6756 kb/s
=> After: 4591 kb/s

(Note: etv5k.y4m is not redistributable but is the same file that was used in
<http://blogs.gnome.org/rbultje/2014/02/22/the-worlds-fastest-vp9-decoder-ffvp9/>)

FIXME: To avoid undershooting(3798 kb/s), the change to inter_minq from
38ff3e328f95c040f96de798d3d4ba3dd510ab43 had to be reverted.

Change-Id: I179f00a9ce77ea5343491f6a1bd61424b60caf81