Update patch set 2
Patch Set 2:

I tested against a video with different levels of artificial noise (gaussian, i.i.d.) added.

This commit, the previous commit, and the program without denoising enabled were compared.

For speed 5, this version strictly beats the previous version, which meets or exceeds no denoising.

For speed 6, the previous version outperforms this one until sigma=3, after which the new version outperforms the previous.

Results are comparable for different bit rates.

One point of interest is that for very high levels of noise, the performance of the denoiser in this version does not tend toward that of no denoising, suggesting that we are denoising too aggressively. Indeed, looking at the logs, we apply the strong denoising filter to >90% of blocks.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0