Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

DO NOT SUBMIT YET

We are currently working on merging the latest origin/master into highbitdepth.  If more changes are made to highbitdepth we will have to restart the merge so please do not submit this until the merge has been submitted and this patch has been rebased.

Patch-set: 1
Label: Commit-Queue=0
Label: Verified=0