Make encoder frame buffer allocation on-demand
Change the encoder buffer handling code to allocate
on-demand.

This changes introduces the concept that the reference
frame map may have "empty" slots, that is there is no
valid reference frame indicated in this position. This
state is signaled in the code by an entry -1 in the
empty slot. All reference map slots are now initialized
to the empty state (-1) and each buffer is initialized
to have a count of 0.

Some of the other frame buffers use throughout the code
could still conditionally allocated. This will be done
in a future patch.

Change-Id: Ib28b65906dc60a91c858b8bb73f3b4b08533b832