Update patch set 2
Patch Set 2:

not my favorite way to fix so I grouped them all.   

I know the preferred method is not to have a default case for an enum rather we should let the compiler catch any unhandled case  ... but the current code has lots of places where the last value in enum is used for a count and quite a few places where an int is cast to an enum value.  As such this is a quicker fix to let us avoid warning errors.

Patch-set: 2
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0