Skip comp inter mode test in RD loop with same frame bias signs
This commit allows the encoder to skip check on compound inter
modes in the rate-distortion optimization loop, if the reference
frame bias signs are the same. Such modes were tested in the
rate-distortion loop to estimate the frame level reference mode,
but were discarded in the block mode selection stage. This commit
makes the encoder to skip such tests in speed 3.

Change-Id: Ib753e6bb11cbdd338aee69dbe2b649671f75a6b0