Add SSE code for block 16x16 in VP9 denoiser.
The SSE code is inherited from VP8 denoiser. In VP8, there are
only 16x16 blocks in denoiser, but in VP9, there are different size
of blocks. For each size, the SSE code will be different in some way.
By adding this SSE code, the speed improvement is around 20%, vary for
different clips.

Change-Id: I03d2c4577afa3d016ad98381936ff1d1c679d3b7