Use C code with problematic compilers
The code in this change is implemented as a workaround for
compilers which can not use vst4_lane_u8. It is not intended
to be performant, only correct.

We have reports of the workaround code crashing on unaligned
access with the vst1_lane_u32 instruction. While it is expected
that the stores are unaligned (the destination is offset by 2
so that the 4 stored bytes cross the block boundery) it is not
clear why only some targets crash with this code since most
ARM instructions can tolerate this misalignment.

The short walsh change is similar in that the neon code generates
an internal compiler error with some versions of gcc4.6

BUG=860
BUG=webrtc:3814

Change-Id: Ide7b747a1201c90c29c112d381d668eb98029cab