Add a level of indirection to loopfilter stores
The code in this change is implemented as a workaround for
compilers which can not use vst4_lane_u8. It is not intended
to be performant, only correct.

We have reports of the workaround code crashing on unaligned
access with the vst1_lane_u32 instruction. This still needs
investigation to determine if the buffer should be aligned or
the code needs to account for unaligned buffers.

In the meantime, it has been changed to use alignment-tolerant
stores.

BUG=webrtc:3823
BUG=webrtc:3814

Change-Id: Ide7b747a1201c90c29c112d381d668eb98029cab