Two pass rc changes.
Adjustments to the GF interval choice and minimum boost.
Adjustment to bits_per_mb estimates to reduce tendancy
to undershoot. Overall improves accuracy of rate control
but hurts metrics a little (derf 0.15%, stdhd 0.3%)

Compared to the September 18 baseline the accuracy of
the VBR rc measured on the derf set is as follows:-

Mean error %  / Mean abs(error %)
Sept 18 baseline (-7.33% / 14.99%)
This check in (-3.19% / 14.18%)

The mean undershoot is reduced slightly but the
worst case overshoot on e.g. harbour/highway is
increased. This will be addressed in a later patch.

Change-Id: Iffd9b0ab7432a131c98fbaaa82d1e5b40be72b58