Add limited adaption of two pass min and max q range.
To improve rate control in VBR case add ability to stretch
the Q range where overshoot or undershoot is building up.

This does tend to hurt quality metrics (especially at very
high rates / lowq) but helps to trim back the minority
of clips which are currently overshooting and boosts
the rate on some that are undershooting.

Change-Id: I422dd45202aa2a6f2614f40f8192e5798b89f895