Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

With the more accurate constraint on gcc4.6, this issue has been alleviated. It occurred because clang was optimizing the store by adding an alignment directive based on the size of the operand. Since we were using this as a hack to force storage across a word boundary, it's not surprising the code crashed.

The older compiler that this originally targeted does not make this optimization and it's unlikely that someone would backport it to gcc4.6.

Leaving this sitting around for a bit but it will likely get abandoned at some point.

Patch-set: 2
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Commit-Queue=0
Label: Verified=0