Update patch set 2
Patch Set 2: Do not submit

Today's adventure in running vpxdec on the ipad was entertaining. Commit message contains the fun bits of the journey to running it...

It does crash on the big buck bunny trailer (after remuxing to IVF). Shell reports "Bus error: 10"

Marking do not submit/-2 for now-- this almost certainly is not something that should be landed, and certainly not in its present state.

Patch-set: 2
Label: Code-Review=-2
Label: Commit-Queue=0
Label: Verified=0