Update patch set 3
Patch Set 3:

Patch 2 failed the CpuSpeedTest(Q=0 case). The reason is that, when var=0 or sse=0, skip_txfm is set to 2 or 1 instead of 0. Then an optimized fast encoding is done, which gives slightly different result. So, the psnr is 88.6354 instead of 100, which causes the failure.

var/sse = 0 should be allowed to do fast encoding. We need to fix the code.

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0