Add adaptation option for VBR.
Allow min and maxQ to creep when the undershoot
or overshoot exceeds thresholds controlled by the
command line under_shoot_pct and over_shoot_pct
values.

Default is 100%,100% which ~disables adaptation.

Derf results for example overshoot / undershoots %:-

Head:- Mean abs (%rate error) = 14.4%

This check in:-
25%/15% - Mean abs (%rate error) = 9%
                  Psnr hit -0.4%

5% / 5%  - Mean abs (%rate error) = 3.5%
                 Psnr hit -2.4%

Most of the remaining error and most of the quality hit is
at extreme data rates. The adaptation code still has an
exception for material that is in effect static so that we
don't over adjust and over spend on YT slide show type
content.

Change-Id: If25a2449a415449c150acff23df713e9598d64c9