Break the encoder to demonstrate segmentation map rules.
When update_map is not set on an inter frame the segmentation map from
the previous frame carries forward, even if segmentation is disabled on
the current frame. If segmentation is disabled on the current frame, all
blocks in the current frame get set to segment zero, but the the
segmentation map remains in place for future frames. The only exception
to this is:
A keyframe, intra-only frame, or frame coded in error resilient mode
always discards the previous segmentation map.

In general think of the segmentation-map state independently from the
segmentation-enabled state. Segmentation can flip on and off without
changing the map, and the map can change whenever segmentation is on.

Change-Id: Ifa4d1dc0a350f498fc2ab3430fc680f3c67bedc2