Allow update of golden refernce buffer in CBR mode
This commit changes to allow the usage of golden reference frame in
VP9 CBR mode to improve quality. VP9 supports potentially up to 8
reference buffers, it has reference buffers available for this
purpose. This was not possible in VP8 as golden and alt-ref buffers
were used for temporal scalability purpose in CBR mode in WebRTC.

For frames that update inter frame, there can be a boost to quality.
The amount of allowed bitrate boost can be controlled via parameter
rc_max_inter_bitrate_pct. (--max-inter-rate for vpxenc)

Visual checking: the change was able to help area covered/unconvered
by moving objecct to recover quality faster at --cpu-used=-5 and -6.

Change-Id: I0c5f010c8fd8b7b598f69779c1b30e5b2ac30a4d