Non-transform coding experiment
In the non-transform coding mode, the prediction error is directly
copied to coefficient buffer for quantization and dequantization.
Accordingly, decoder adds the de-quantized coefficients directly
to its predictions to obtain reconstructed pixel values.

Encoder decides whether to use the non-transform coding mode based
on RD comparison. 2 flags are added to the MD ModeInfo structure
to indicate if transform is skipped for Y and UV channel, respec-
tively.

The experiment is enabled by setting CONFIG_TX_SKIP to be 1.
Currently, non-transform option is only available for keyframes.

Change-Id: I6e48285037ac1e3ef8bd72f53006bdea7d7a883e