Non-transform coding experiment
In the non-transform coding mode, the prediction error is directly
copied to coefficient buffer for quantization and dequantization.
Accordingly, decoder adds the de-quantized coefficients directly
to its predictions to obtain reconstructed pixel values.

Decoder decides whether to use the non-transform coding mode based
on RD comparison. 2 flags are added to the MD modeinfo structure
to indicate if transform is skipped for Y and UV channel, respec-
tively.

The experiment is enabled by setting CONFIG_TX_SKIP to be 1.
Currently, it only works for keyframes, and force partition size
to be larger than 4X4. To be continued...

Change-Id: I6e48285037ac1e3ef8bd72f53006bdea7d7a883e