Non-transform coding experiment
In the non-transform coding mode, the prediction error is directly
copied to coefficient buffer for quantization and dequantization.
Accordingly, decoder adds the de-quantized coefficients directly
to its predictions to obtain reconstructed pixel values.

Encoder decides whether to use the non-transform coding. Ideally,
this should be done based on RD comparison. But currently, the
non-transform coding mode is selected randomly. 2 flags are added
to the MD ModeInfo structure to indicate if transform is skipped
for Y and UV channel, respectively.

The experiment is enabled by setting CONFIG_TX_SKIP to be 1.

Change-Id: I6e48285037ac1e3ef8bd72f53006bdea7d7a883e